Add NTT computations and use them in TFHE
Doesn’t work yet, since TLweFFTAddMulRTo (which operates in the finite field space) still uses regular integer arithmetic. To be fixed.

+++ tfhe/numeric_functions.py
@@ -2,7 +2,7 @@ import numpy
 
 
 Torus32 = numpy.int32
-Complex = numpy.complex128
+Complex = numpy.uint64
 Float = numpy.float64
